requirements.txt
groq==0.9.0
chromadb==0.5.3
sentence-transformers==2.3.1
streamlit==1.40.1
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
